+++ tonlibjson-client/src/client.rs
@@ -140,7 +140,12 @@ pub struct ResponseFuture<R> {
 #[pinned_drop]
 impl<R> PinnedDrop for ResponseFuture<R> {
     fn drop(self: Pin<&mut Self>) {
-        if let ResponseState::Rx { request_id, request_storage, ..} = &self.state {
+        if let ResponseState::Rx {
+            request_id,
+            request_storage,
+            ..
+        } = &self.state
+        {
             request_storage.remove(request_id);
         }
     }
@@ -181,12 +186,12 @@ where
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let mut this = self.project();
 
-        return match this.state.as_mut().project() {
+        match this.state.as_mut().project() {
             ResponseStateProj::Failed { error } => {
                 Poll::Ready(Err(error.take().expect("polled after error")))
             }
             ResponseStateProj::Rx { rx, .. } => {
-                return match ready!(rx.poll(cx)) {
+                match ready!(rx.poll(cx)) {
                     Ok(response) => {
                         // TODO[akostylev0] refac!!
                         if response.data["@type"] == "error" {
@@ -205,9 +210,9 @@ where
                         }
                     }
                     Err(_) => Poll::Ready(Err(anyhow!("oneshot closed"))),
-                };
+                }
             }
-        };
+        }
     }
 }